Data Analytics and Exploration/Healthcare RAG/requirements.txt
pandas
langchain
transformers
autoawq